requirements.txt
# FastAPI ve ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Veritabanı
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Güvenlik
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validasyon
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1

# E-posta
aiosmtplib==3.0.1

# Diğer
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1
